create / implement SkillRow component
